backend/requirements.txt
Django==3.2.16
django_colorfield==0.8.0
django_filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf_extra_fields==3.4.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.8.6
python-dotenv==0.20.0
reportlab==3.6.12
django-cors-headers==3.7.0
